feat: adding function for sum values
