+++ lib/parser.ex
@@ -1,0 +1,14 @@
+defmodule ReportsGenerator.Parser do
+  def build(filename) do
+    "reports/#{filename}"
+    |> File.stream!()
+    |> Stream.map(fn line -> parse_line(line) end)
+  end
+
+  def parse_line(line) do
+    line
+    |> String.trim()
+    |> String.split(",")
+    |> List.update_at(2, &String.to_integer/1)
+  end
+end
