+++ lib/reports_generator.ex
@@ -1,14 +1,15 @@
 defmodule ReportsGenerator do
+  alias ReportsGenerator.Parser
+
   def build(filename) do
-    "reports/#{filename}"
-    |> File.stream!()
-    |> Enum.map(&parse_line(&1))
+    filename
+    |> Parser.parse_line()
+    |> Enum.reduce(report_acc(), fn line, report ->
+      sum_values(line, report)
+    end)
   end
 
-  def parse_line(line) do
-    line
-    |> String.trim()
-    |> String.split(",")
-    |> List.update_at(2, &String.to_integer/1)
-  end
+  defp sum_values([id, _food, price], report), do: Map.put(report, id, report[id] + price)
+
+  defp report_acc, do: Enum.into(1..30, %{}, &{Integer.to_string(&1), 0})
 end
